// site/src/types/API.ts
export enum API {
    NETWORK_SET = '/api/network/set',
    NETWORK_GET = '/api/network/get',

    UPLOAD_POINTS = '/api/upload',
    DOWNLOAD_POINTS = '/api/points',

    STOP = '/api/signal/stop',
    START = '/api/signal/start',
    PAUSE = '/api/signal/pause',
    CLEAR = '/api/signal/clear',
    REBOOT = '/api/signal/reboot',
}